PythonSoftwareProjects/AiTwitterLabelingMicroservice/requirements.txt
psycopg
tiktoken
numpy
redis
pandas
requests
python-dotenv
openai
pika